flask-uploadfile/requirements.txt
Flask (0.10.1)
Flask-RESTful (0.3.5)
mysql-connector-python (1.1.6)
requests (2.7.0)

